requirements.txt
Django==2.0.7
django-webpack-loader==0.6.0
djangorestframework==3.8.2
Pillow==5.2.0
pytz==2018.5
